Split project into submodules to allow better sharing of test helpers and WS code.
