+++ core/src/main/scala/me/jeffmay/neo4j/client/Neo4jClient.scala
