+++ core/src/main/scala/me/jeffmay/neo4j/client/StatusCodes.scala
