+++ core/src/main/scala/me/jeffmay/neo4j/client/cypher/package.scala
