+++ core/src/main/scala/me/jeffmay/neo4j/client/errors/exceptions.scala
