+++ core/src/main/scala/me/jeffmay/neo4j/client/scalacheck/package.scala
