+++ core/src/main/scala/me/jeffmay/util/Classes.scala
