+++ ws-util/src/main/scala/me/jeffmay/util/concurrent/FutureWithTimeout.scala
