+++ ws/src/main/scala/me/jeffmay/neo4j/client/rest/package.scala
