+++ ws/src/main/scala/me/jeffmay/neo4j/client/ws/WSRequestWithException.scala
