+++ ws/src/test/scala/me/jeffmay/neo4j/client/ws/WSNeo4jClientBasicSpecs.scala
